+++ client/src/scripts/news/new.ts
@@ -1,6 +1,66 @@
+import { formatToHtml } from "../engine/utils.ts";
+
 const newsS=document.querySelector("#about-sm-news") as HTMLDivElement
 
 const news=[
+    {
+      title: "Pre-Beta-2",
+      content:`
+## News
+___
+### Mains
+* **SNOW MAP!!** Mary Christmas
+* Containers
+* Christmas Tree
+* A New Movimentation
+* Snow Storm
+* Ice
+* Ice Slide
+* Biome Musics
+* Day And Night Cycle
+___
+### Menu
+* Menu Music
+* Loading Screen
+* A Sound For Buttons Click
+* Normal Mode Image Changed To A Gif
+* A New Gif For Snow Mode
+* Accounts Enabled For Test - PLEASE DONT USE YOUR REAL PASSWORD
+* News Now Suport Markdown
+___
+### Items
+* Yellow Laser Pistol: A Weapon Madded To Test Reflections
+* Red Soda: A Alternative For Small Red Crystal
+* M9 Sounds
+* M9 Placeholder Graphics
+* Green Soda And Green Bless
+* Black Soda And Death
+* Medikit Sfx
+* Bandage Sfx
+* Pills Sfx
+___
+### Others
+* New Snow Musics
+___
+## Changes
+___
+### Items
+* Quickswitch Improvement
+* Survival Knife Attack Delay Come From 0.2s To 0.4s
+* Bandage Now Healing Come From 15 To 20
+* Soda Become Yellow Soda
+* Small Blue Potion Become Blue Soda
+* Small Purple Potion Become Purple Soda
+* Small Red Crystal Just Heal Until 50%
+___
+### Others
+* Readd Rivers
+* Some Musics Change The Biome
+___
+## Fixes
+* Rect Hitbox Damage
+      `
+    },
     {
       title: "Pre-Beta",
       content:`
@@ -357,6 +417,6 @@ const news=[
 for(const n of news){
     const d=document.createElement("div")
     d.classList.add("update-item")
-    d.innerHTML=`<h2>${n.title}</h2>`+n.content
+    d.innerHTML=`<h2>${n.title}</h2>`+formatToHtml(n.content)
     newsS.appendChild(d)
 }